requirements.txt
torch>=2.1.0
torchvision
torchaudio
sentencepiece>=0.1.99
datasets>=2.14.0
transformers>=4.35.0
openai-whisper>=20231117
TTS>=0.22.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sounddevice>=0.4.6
numpy>=1.24.0
tqdm>=4.66.0
wandb>=0.16.0
pyyaml>=6.0
jiwer>=3.0.3
scipy>=1.11.0
librosa>=0.10.1
accelerate>=0.24.0
evaluate>=0.4.0
python-multipart>=0.0.6
